Speed up the Wormhole vision tower: SDPA, redundant ops, and row padding.
Three passes on top of the matmul program configs, measured at depth 1 on the
demo grid (patch grid 1x86x128 = 11008 patches). 27-block tower device time:

  9B  / N300  934.0 -> 731.4 ms (1.28x), full-depth PCC 0.98540 -> 0.99921
  27B / T3K   906.4 -> 756.6 ms (1.20x), full-depth PCC 0.98696 -> 0.99897

SDPA. Fidelity was HiFi4 and K was bf16 while Q/V were bf8b, both chosen for
operands the tower does not use -- `kv_cache_dtype` never applied here, this is
single-pass non-causal attention with no KV cache, and it also left a bf16->bf16
no-op typecast on the profile. HiFi2 + bf8b K, and chunk sizes swept per mesh.
LoFi rejected at PCC 0.9656/op; 512/512 rejected, its flash CBs need 1,949,888 B
against L1's 1,499,136.  18,119 -> 14,430 us/block (9B), 6,103 -> 4,265 (27B).

Redundant data movement. `rot_mats` were padded 72->96 inside the attention, so
the same two tensors were re-padded once per block, 27x per image; hoisted to the
host, where the in-attention pad now gates on the tensor's own last dim so other
callers still work. pixel_values, the bilinear weights and cos/sin were uploaded
fp32, so ttnn tilized fp32 on device and then typecast; cast on host instead,
which also halves the bytes over PCIe. FillPad 6 -> 0, TM -34%/-37%.

Row padding. `seq_len` rounded to a multiple of 2048, justified in a comment as
"required by tt_transformers.tt.attention::forward_prefill" -- a file this tower
never calls, it has its own VisionAttention. The real constraints are
`seq_len % 128 == 0`, matmul chunks that are already derived as divisors of the
row count, and ttnn SDPA, which validates only `chunk % 32` and handles a ragged
tail. So the demo grid needs no padding at all: 11008 = 86 x 128. Also repairs an
off-by-one -- `(n // m) + 1` over-padded exact multiples, sending a 4096-patch
image to 6144, i.e. 2.25x the SDPA for nothing.

This one is a correctness fix as much as a perf fix. SDPA runs is_causal=False
with no attn_mask, so the pad rows were unmasked *keys*: zeros, so every real
query scored q.0 = 0 against each of the 1280 of them and summed exp(0) = 1 into
its softmax denominator. Removing them is worth 18x on the 9B's full-depth PCC.
An earlier note in VISION_TOWER_PERF.md blamed the ~0.985 on bfloat8_b weight
error and called it pre-existing; that was wrong, and the control that suggested
it (QWEN36_VISION_MM_TUNING=0, which measured no better) could not see it because
the env gate does not touch seq_len -- both arms carried the same bug.

Changing the row count invalidated the chunk sweep, so it was re-run: both meshes
now want q=128/k=512 and _VISION_SDPA_TUNING_BY_DEVICE is empty. The old
rationale ("winning q_chunk depends on head count") was a two-point fit; what
actually mattered is that 8 x (12288/256) = 384 is exactly 6 rounds of 64 cores,
which stops holding the moment the rows are not a multiple of 256*64/heads.

All of it stays behind `vision_mm_tuned` (Wormhole only, QWEN36_VISION_MM_TUNING
to disable) except the redundant-op and padding fixes, which are unconditional.
No Blackhole card was available, so P150/P150x4 keep the untuned graph.

Verified on both meshes: depth-1 profile, full-depth PCC, isolated SDPA sweep,
and the traced_single_image demo (9B TTFT 2.234 s / 27.7 tok/s, 27B 2.707 s /
18.9 tok/s, both content-checked).

+++ models/demos/blackhole/qwen36/tt/vision/patch_embed.py
@@ -159,7 +159,10 @@ class VisionEmbed(LightweightModule):
         rows = ((n + 31) // 32) * 32
 
         # --- patch projection ---
-        x = pixel_values
+        # Cast on host: the processor hands us fp32, and uploading that made ttnn tilize fp32 on
+        # device and then typecast (849 + 481 us/image on an N300). bf16 halves the bytes over PCIe
+        # and lands in the dtype the matmul wants.
+        x = pixel_values.to(torch.bfloat16)
         if rows != n:
             x = torch.nn.functional.pad(x, (0, 0, 0, rows - n))
         x_tt = ttnn.from_torch(
@@ -195,7 +198,7 @@ class VisionEmbed(LightweightModule):
 
         # --- interpolated positional embedding: sum_c w_c * table[idx_c] ---
         idx = bilinear_indices.to(torch.int32)
-        wts = bilinear_weights.to(torch.float32)
+        wts = bilinear_weights.to(torch.bfloat16)  # same reason as the pixels above
         if rows != n:
             # Pad rows look up entry 0 with weight 0, contributing exactly nothing.
             idx = torch.nn.functional.pad(idx, (0, rows - n))
